% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_mean_map.R
\name{laplacian}
\alias{laplacian}
\title{Compute the Laplacian matrix for a model.frame}
\usage{
laplacian(data_bags, feature_mat, bag_x_avg, similarity = "G,s",
  sigma = 10, epsilon = 0, ...)
}
\arguments{
\item{mf}{a model.frame object}
}
\description{
Compute the Laplacian matrix for a model.frame
}
