% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_label_proportions.R
\name{llp}
\alias{llp}
\title{Learning from Label Proportions}
\usage{
llp(formula, data, bag_proportions, mode = "LLM", alternating = T, ...)
}
\arguments{
\item{formula}{A formula object describing the bag vs.
features to use.}

\item{data}{An R data.frame where each row represents a
training instance. Must contain a bag column,
referred to in the formulat.}

\item{bag_proportions}{A lookup of bag id to the proportion of the
bag in the class}

\item{mode}{The starting mode for the mean map generation}

\item{alternating}{Whether to continue once the first optimisation
is complete using the alternating mean map method.}

\item{...}{Arguments to be passed to downstream components.
See, laplacian, laplacian_mean_map, and
optimise_alternating for possible values.}
}
\description{
Implements "(Almost) No Label No Cry", learning
from label proportions algorithms. Including
laplacian mean map, mean map, and alternating
mean map methods.
}
