% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_map.R
\name{mean_map}
\alias{mean_map}
\title{Mean Map algorithm}
\usage{
mean_map(data_bags, feature_mat, bag_proportions, weight = NULL, ...)
}
\arguments{
\item{weight}{is matrix D_w in the paper}
}
\description{
The algorithm expects the data organized in data$bag and data$label (the proportions).
The rest is the features vector
}
