% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternating.R
\name{optimise_alternating}
\alias{optimise_alternating}
\title{Optimise with the Alternating Mean Map method}
\usage{
optimise_alternating(data_bags, feature_mat, proportions, model,
  minmax = FALSE, maxcount = 30, ERR = 0.001)
}
\arguments{
\item{data_bags}{A factor vector of the bag factor for the training dataset.
Filled in by llp}

\item{feature_mat}{A matrix of features for each training instance
Filled in by llp}

\item{model}{Initial model to use, as a weights vector for the features.}

\item{minmax}{Whether to use the AMM^max (when TRUE), or the default AMM^min
algorithm.}

\item{maxcount}{Maximum number of iterations to use under the AMM^max algorithm}

\item{ERR}{Stopping condition for the change in the loss in each round}

\item{bag_proportions}{A numerical vector, with each entry representing the proportion
of positive instances for each bag. Must be in the order of the
levels of the data_bags factor.
Filled in by llp}
}
\description{
Performs logistic regression in an optimisation
loop, where the positive instances in each bag
are determined from the previous iteration.
}
\details{
One shouldn't need to call this function directly, as the `llp`
  function will invoke it with the correct parameters.
}
