% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_shot.R
\name{optimise_one_shot}
\alias{optimise_one_shot}
\title{Optimise a linear learner using the mean operator}
\usage{
optimise_one_shot(mat, mean_operator, weight = NULL)
}
\arguments{
\item{mat}{a matrix}

\item{mean_operator}{a mean.operator object}

\item{weight}{an optional vector of weights to be used in the fitting
process. Should be ‘NULL’ or a numeric vector.}
}
\description{
Given a true mean operator, this will yield near
identical results to glm with binomial(link = "logit").
}
