% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_label_proportions.R
\name{oracle}
\alias{oracle}
\title{Logistic regression}
\usage{
oracle(formula, data)
}
\arguments{
\item{formula}{A formula object describing the bag vs.
features to use. The reponse variable will
be coerced to a `factor` if it is not already.}

\item{data}{An R data.frame where each row represents a
training instance. Must contain a bag column,
referred to in the formula.}
}
\description{
But formulated within the llp framework.
}
