% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_mean_map.R
\name{laplacian}
\alias{laplacian}
\alias{laplacian_mean_map}
\title{Laplacian Mean Map}
\usage{
laplacian(data_bags, feature_mat, bag_x_avg, similarity = "G,s",
  sigma = 10, epsilon = 0, ...)

laplacian_mean_map(data_bags, feature_mat, bag_proportions, gamma = 1,
  weight = NULL, ...)
}
\arguments{
\item{data_bags}{A factor vector of the bag factor for the training dataset.
Filled in by llp}

\item{feature_mat}{A matrix of features for each training instance
Filled in by llp}

\item{bag_x_avg}{Feature averages, by bag. Filled in by laplacian_mean_map}

\item{similarity}{Similarity function for determing the laplactian. Available
options are "G,s" and "NC".}

\item{sigma}{Scaling constant used in the G,s algorithm}

\item{epsilon}{Small constant to add to diagonal elements of the laplacian.
Values other than 0 mean the returned matrix will not be a true
laplacian matrix.}

\item{...}{Parameters to pass through to laplacian}

\item{bag_proportions}{A numerical vector, with each entry representing the proportion
of positive instances for each bag. Must be in the order of the
levels of the data_bags factor.
Filled in by llp}

\item{gamma}{Interaction constant for the laplacian.}

\item{weight}{an optional vector of weights to be used in the fitting
process. Should be ‘NULL’ or a numeric vector.}
}
\description{
Generate a laplacian mean map for a dataset.
}
