% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_shot.R
\name{optimise_one_shot}
\alias{optimise_one_shot}
\title{Optimise a linear learner using the mean operator}
\usage{
optimise_one_shot(mf, mean_operator)
}
\arguments{
\item{mf}{model.frame object}

\item{mean_operator}{a mean.operator object}
}
\description{
Optimise a linear learner using the mean operator
}
